ufx_devel/requirements.txt
Django==1.11.4
pytz==2017.2
psycopg2
djangorestframework
gunicorn==19.3.0
django-storages
django-rest-swagger
